# apidog-import-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Feedbase Backend API (APIdog Import)",
    "version": "1.0.0",
    "description": "APIdog import file generated from API_FULL_LIST.md."
  },
  "servers": [
    {
      "url": "http://localhost:4560",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/votes/add": {
      "post": {
        "tags": [
          "votes"
        ],
        "summary": "POST /votes/add",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "message": "Vote added successfully"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lg": {
                    "type": "string"
                  },
                  "postId": {
                    "type": "integer"
                  }
                }
              },
              "example": {
                "lg": "en",
                "postId": 101
              }
            }
          }
        }
      }
    },
    "/votes/remove/{postId}": {
      "delete": {
        "tags": [
          "votes"
        ],
        "summary": "DELETE /votes/remove/:postId",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "message": "Vote removed successfully"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lg": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "lg": "en"
              }
            }
          }
        }
      }
    },
    "/votes/post/{postId}": {
      "post": {
        "tags": [
          "votes"
        ],
        "summary": "POST /votes/post/:postId",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "message": "Votes retrieved successfully",
                  "data": [
                    {
                      "id": 301,
                      "post_id": 101,
                      "user_id": 3,
                      "vote_type": "upvote",
                      "full_name": "Jane Product",
                      "email": "jane@acme.test"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lg": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "lg": "en"
              }
            }
          }
        }
      }
    },
...
  }
}